requirements.txt
streamlit
numpy
pandas
scikit-learn
pillow
mne      
nibabel            
pydicom           
torch
torchvision
pydicom
nibabel
opencv-python
gradio
albumentations
transformers
huggingface-hub
tensorflow
google-generativeai